+++ src/main/ruby/truffleruby/core/rationalizer.rb
@@ -79,8 +79,8 @@ class String
       end
     end
 
-    def convert
-      internal_convert() || Rational(0, 1)
+    def convert(default = Rational(0, 1))
+      internal_convert() || default
     end
 
     def strict_convert